requirements.txt
python-decouple
https://github.com/New-dev0/Telethon/archive/diamond.zip